requirements.txt
torch
torchvision
torchaudio
streamlit
pillow
matplotlib
scikit-learn
seaborn
